Extract Redux internals caches and nuke dead/WIP code
